+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Main (unreleased)
 
 - New Grafana Agent Flow components:
 
-  - `otelcol.processor.spanlogs` - creates logs from spans. It is the flow mode equivalent 
+  - `otelcol.connector.spanlogs` - creates logs from spans. It is the flow mode equivalent 
   to static mode's `automatic_logging` processor. (@ptodev)
 
 v0.36.0-rc.0 (2023-08-25)